refactor test
add 'testcontainers-annotations', migrate on JUnit 5
