+++ src/test/java/com/rbkmoney/payout/manager/dao/PayoutDaoTest.java
@@ -6,7 +6,7 @@ import com.rbkmoney.payout.manager.domain.tables.pojos.Payout;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import static com.rbkmoney.payout.manager.config.AbstractDaoConfig.generatePayoutId;
+import static com.rbkmoney.payout.manager.util.ValuesGenerator.generatePayoutId;
 import static com.rbkmoney.testcontainers.annotations.util.RandomBeans.random;
 import static org.junit.jupiter.api.Assertions.*;
