+++ src/test/java/com/rbkmoney/payout/manager/util/ValuesGenerator.java
@@ -1,0 +1,10 @@
+package com.rbkmoney.payout.manager.util;
+
+import java.util.UUID;
+
+public class ValuesGenerator {
+
+    public static String generatePayoutId() {
+        return UUID.randomUUID().toString();
+    }
+}
